font moved, mostly updated
