+++ meerk40t/gui/plugin.py
@@ -44,6 +44,7 @@ def plugin(kernel, lifecycle):
     if lifecycle == "preregister":
         from meerk40t.gui.laserrender import LaserRender
         from meerk40t.gui.wxmeerk40t import wxMeerK40t
+        from meerk40t.gui.fonts import svgfont_to_wx, wxfont_to_svg
 
         kernel.register("module/wxMeerK40t", wxMeerK40t)
         kernel_root.open("module/wxMeerK40t")
@@ -51,6 +52,8 @@ def plugin(kernel, lifecycle):
         # Registers the render-op make_raster. This is used to do cut planning.
         renderer = LaserRender(kernel_root)
         kernel_root.register("render-op/make_raster", renderer.make_raster)
+        kernel_root.register("font/svg_to_wx", svgfont_to_wx)
+        kernel_root.register("font/wx_to_svg", wxfont_to_svg)
     if lifecycle == "register":
 
         from meerk40t.gui.scene.scene import Scene
