+++ meerk40t/gui/propertypanels/textproperty.py
@@ -1,6 +1,6 @@
 import wx
 
-from meerk40t.core.fonts import wxfont_to_svg
+from meerk40t.gui.fonts import wxfont_to_svg
 
 from ...svgelements import Color
 from ..icons import icons8_choose_font_50, icons8_text_50
